backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.30
pymysql>=1.1.1
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
faiss-cpu>=1.8.0
numpy>=2.0.0
